ignore tfjob/pytorch job if corresponding CRD not created
