+++ pkg/controller/studyjob/studyjob_controller.go
@@ -51,6 +51,10 @@ const (
 	cleanDataFinalizer = "clean-studyjob-data"
 )
 
+var (
+	invalidCRDResources [] string
+)
+
 /**
 * USER ACTION REQUIRED: This is a scaffold file intended for the user to modify with their own Controller
 * business logic.  Delete these comments after modifying this file.*
@@ -117,7 +121,8 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 			OwnerType:    &katibv1alpha1.StudyJob{},
 		})
 	if err != nil {
-		return err
+		invalidCRDResources = append(invalidCRDResources, TFJobWorker)
+		log.Printf("Fail to watch TFJob resource: %v", err)
 	}
 
 	err = c.Watch(
@@ -127,7 +132,8 @@ func add(mgr manager.Manager, r reconcile.Reconciler) error {
 			OwnerType:    &katibv1alpha1.StudyJob{},
 		})
 	if err != nil {
-		return err
+		invalidCRDResources = append(invalidCRDResources, PyTorchJobWorker)
+		log.Printf("Fail to watch PyTorchJob resource: %v", err)
 	}
 
 	return nil
